\name{rSCR-package}
\alias{rSCR-package}
\alias{rSCR}
\docType{package}
\title{
Functions to interact with Synapse Commons Repository (SCR)
}
\description{
Useful functions for getting data into and out of the Synapse Commons Repository.
}
\details{
\tabular{ll}{
Package: \tab rSCR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-03\cr
License: \tab GNU\cr
}
}
\author{
Brig Mecham
}
\references{
}
\keyword{ package }
\seealso{

}
\examples{
a <- 10

}
